# openapi.yaml
openapi: 3.0.3
info:
  title: Midas DVR / KLV API
  version: 1.0.0
  description: |
    HTTP API for source ingest, authoritative file workflows, HLS/KLV state,
    clips, snapshots, telemetry, and WebRTC signaling. File-backed operations
    use the private authoritative asset stored under mission-products/<productId>/source.
servers:
  - url: http://localhost:8090
    description: Default local server
...
paths:
...
  /streams/{streamId}/klv:
    get:
      tags: [Telemetry]
      summary: Query stored KLV events by timestamp range
      parameters:
        - { $ref: '#/components/parameters/StreamId' }
        - { name: fromMs, in: query, required: true, schema: { type: number } }
        - { name: toMs, in: query, required: true, schema: { type: number } }
      responses:
        '200': { description: KLV event collection }
        '400': { $ref: '#/components/responses/Error' }
  /streams/{streamId}/klv/platform-history.geojson:
    get:
      tags: [Telemetry]
      summary: Get compact sampled platform history as GeoJSON
      description: Returns a single GeoJSON LineString sampled from the last valid platform position in each completed browser HLS segment. Coordinates remain in HLS sequence order; `properties.timesMs` is index-aligned with those coordinates for mission-time filtering. Consecutive duplicate positions are removed and very long paths are bounded while preserving first and last points. This compact map index does not replace full-rate KLV event storage.
      parameters:
        - { $ref: '#/components/parameters/StreamId' }
        - { name: fromMs, in: query, required: false, schema: { type: number }, description: Optional inclusive mission-time start in Unix milliseconds. }
        - { name: toMs, in: query, required: false, schema: { type: number }, description: Optional inclusive mission-time end in Unix milliseconds. }
        - { name: maxPoints, in: query, required: false, schema: { type: integer, minimum: 2, maximum: 10000 }, description: Maximum returned path coordinates, subject to the server limit. }
      responses:
        '200':
          description: GeoJSON Feature with a LineString geometry when at least two platform points are available
          content:
            application/geo+json:
              schema: { type: object }
        '400': { $ref: '#/components/responses/Error' }
  /streams/{streamId}/klv/frame-center-history.geojson:
    get:
      tags: [Telemetry]
      summary: Get compact sampled frame-center history as GeoJSON
      description: Returns a single GeoJSON LineString sampled from the last valid frame-center position in each completed browser HLS segment. Coordinates remain in HLS sequence order; `properties.timesMs` is index-aligned with those coordinates for mission-time filtering. Consecutive duplicate positions are removed and very long paths are bounded while preserving first and last points. This compact map index does not replace full-rate KLV event storage.
      parameters:
        - { $ref: '#/components/parameters/StreamId' }
        - { name: fromMs, in: query, required: false, schema: { type: number }, description: Optional inclusive mission-time start in Unix milliseconds. }
        - { name: toMs, in: query, required: false, schema: { type: number }, description: Optional inclusive mission-time end in Unix milliseconds. }
        - { name: maxPoints, in: query, required: false, schema: { type: integer, minimum: 2, maximum: 10000 }, description: Maximum returned path coordinates, subject to the server limit. }
      responses:
        '200':
          description: GeoJSON Feature with a LineString geometry when at least two frame-center points are available
          content:
            application/geo+json:
              schema: { type: object }
        '400': { $ref: '#/components/responses/Error' }
  /streams/{streamId}/klv/export.csv:
    get:
      tags: [Telemetry]
      summary: Download all stored KLV telemetry for a stream as CSV
      description: One row per decoded KLV event, ordered by mission timestamp. The export includes normalized telemetry columns; it does not include Mission Target Log entries.
      parameters: [ { $ref: '#/components/parameters/StreamId' } ]
      responses:
        '200':
          description: UTF-8 CSV telemetry download
          content:
            text/csv:
              schema: { type: string, format: binary }
        '400': { $ref: '#/components/responses/Error' }
  /streams/{streamId}/klv/export.kml:
    get:
      tags: [Telemetry]
      summary: Download platform, SPI, and target KLV time series as KML
      description: Creates compact Google Earth gx:Track time series from stored SQLite KLV telemetry. Platform positions use sensor latitude/longitude/altitude; Sensor - Frame Center positions use frame-center latitude/longitude/elevation; and Target positions use target latitude/longitude/elevation. FOV footprints are omitted. The Platform and SPI tracks contain sample-aligned ExtendedData arrays for UTC mission time, gimbal orientation/FOV, slant range, target size/location, tracking-gate size, and target CE90/LE90. The gx:Track representation is intended for Google Earth rather than per-sample Google Maps popups.
      parameters: [ { $ref: '#/components/parameters/StreamId' } ]
      responses:
        '200':
          description: UTF-8 KML telemetry download
          content:
            application/vnd.google-earth.kml+xml:
              schema: { type: string, format: binary }
        '400': { $ref: '#/components/responses/Error' }
...
components:
  parameters:
    StreamId:
      name: streamId
      in: path
      required: true
      schema: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]{0,127}$' }
...
  responses:
    Error:
      description: Request could not be completed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties: { ok: { type: boolean, example: false }, error: { type: string } }
...